requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
sse-starlette==2.1.3
